Move CI off Node 20, and make the lint that CI runs actually pass
golangci-lint-action v6 runs on Node 20, which GitHub Actions has deprecated. v7 and
later require golangci-lint v2, so .golangci.yml is migrated to the v2 schema — with the
explanatory comments hand-restored, because the reasons for each exclusion are the part
worth keeping. checkout, setup-go and upload-artifact move to their Node 24 majors too.

setup-go pinned go-version: '1.22' while go.mod requires 1.25, so every job bootstrapped
a second toolchain before it could compile anything, and bumping go.mod would not have
updated it. It reads go-version-file: go.mod now.

Running the lint properly surfaced 185 findings. 148 were govet/shadow, every one of them
`err` shadowed by an idiomatic `if err := f(); err != nil` — the check cannot be told to
ignore that, and burying the printf findings under it costs more than it is worth, so
shadow is off with the reasoning recorded. Of the rest:

  - Two guards failed *open*. ConflictingServerName and SudoGrants swallowed a read
    error and returned "no conflict" / "no grants", and both callers then tested
    `err == nil && conflict != ""` — so an unreadable sites-enabled silently opened the
    gate that exists because nginx resolves a duplicate server_name unpredictably. A
    missing directory is still "nothing to conflict with"; anything else is now an
    error, and preflight reports that it could not check rather than claiming the names
    are clear.
  - strings_Dir was a hand-rolled filepath.Dir that returned "." where the real one
    returns "/".
  - Two dead helpers, deleted: an unused function is a claim that something uses it.
  - Sentinel errors compared with == and an error type-asserted, all of which break the
    first time anything wraps them.
  - The Makefile's install hint pointed at the v1 module path, which cannot read this
    config — so a contributor following it would get a linter that refuses to start.

And `ratline help <cmd>` refused with "must run as root". `<cmd> --help` short circuits
inside cobra, but help is a *command*, so the spelling people reach for before they know
the flag existed was the one that would not answer.

The CI smoke test now renders help for every command and every embedded explain topic —
which is the only automated exercise backup, status, troubleshoot, explain and update
get, since the integration suite reaches none of them.

make integration: 137 passed, 0 failed, built from the Dockerfile.

+++ internal/system/fileops_unix.go
@@ -21,5 +21,5 @@ func FreeBytes(path string) (uint64, error) {
 	if err := syscall.Statfs(path, &st); err != nil {
 		return 0, rlerr.Wrap(err, rlerr.CodeGeneric, "checking free space on %s", path)
 	}
-	return uint64(st.Bavail) * uint64(st.Bsize), nil
+	return st.Bavail * uint64(st.Bsize), nil
 }
